Misc: move tokens to special folder and introduce new module
Ref #1910
