+++ lib/errors.js
@@ -1,6 +1,6 @@
 var assert = require('assert');
 var chalk = require('chalk');
-var TokenAssert = require('./token-assert');
+var TokenAssert = require('./tokens/assert');
 
 /**
  * Set of errors for specified file.
