+++ lib/rules/require-spaces-in-for-statement.js
@@ -41,7 +41,7 @@
  */
 
 var assert = require('assert');
-var TokenCategorizer = require('../token-categorizer');
+var TokenCategorizer = require('../tokens/categorizer');
 
 module.exports = function() {};
