+++ lib/rules/require-spaces-inside-parentheses.js
@@ -65,7 +65,7 @@
  */
 
 var assert = require('assert');
-var TokenCategorizer = require('../token-categorizer');
+var TokenCategorizer = require('../tokens/categorizer');
 
 module.exports = function() {};
