+++ lib/tokens/assert.js
