+++ lib/tokens/categorizer.js
