+++ lib/tokens/whitespace.js
@@ -1,0 +1,3 @@
+function WhitespaceBefore() {
+
+}
